Separate rnn code
